Rename common -> Common to match main Dolphin repo
