+++ Common/BitUtils.h
