+++ Common/CommonFuncs.h
