+++ Common/TypeUtils.h
