+++ gxtest/intensity.cpp
@@ -7,7 +7,7 @@
 
 #include <ogcsys.h>
 #include <wiiuse/wpad.h>
-#include "common/hwtests.h"
+#include "Common/hwtests.h"
 #include "gxtest/cgx.h"
 #include "gxtest/cgx_defaults.h"
 #include "gxtest/util.h"
